requirements.txt
anyjson==0.3.3
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
dj-database-url==0.5.0
Django==3.1.2
django-appconf==1.0.4
django-cors-headers==3.2.0
django-extensions==2.2.9
django-extra-fields==2.0.1
djangorestframework==3.11.0
graphviz==0.14
gunicorn==20.0.0
importlib-metadata==1.6.0
kombu==4.6.8
mock==3.0.5
Pillow==6.2.0
psycopg2==2.8.3
pycparser==2.19
pydot==1.4.1
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.3
pytz==2019.3
redis==3.4.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
whitenoise==4.1.4
zipp==3.1.0
